Connection features created
The entire Electron project was organized and new features were defined. The design was made and navigation between pages was done. The new SSH connection saving feature and Passphrase entry for the server we are connecting to were added. All of these were tested and the connection was established properly.